+++ index.js
@@ -33,6 +33,7 @@ const {getAutopilot} = require('./lnd_methods');
 const {getBackup} = require('./lnd_methods');
 const {getBackups} = require('./lnd_methods');
 const {getBlock} = require('./lnd_methods');
+const {getChainAddresses} = require('./lnd_methods');
 const {getChainBalance} = require('./lnd_methods');
 const {getChainFeeEstimate} = require('./lnd_methods');
 const {getChainFeeRate} = require('./lnd_methods');
@@ -184,6 +185,7 @@ module.exports = {
   getBackup,
   getBackups,
   getBlock,
+  getChainAddresses,
   getChainBalance,
   getChainFeeEstimate,
   getChainFeeRate,